+++ tests/testthat/test-scales_fill_hypso.R
@@ -1,3 +1,46 @@
+test_that("Discrete scale", {
+  d <- data.frame(x = 1:5, y = 1:5, z = 21:25, l = letters[1:5])
+
+  p <- ggplot2::ggplot(d) +
+    ggplot2::geom_col(aes(x, y, fill = l))
+
+  init <- ggplot2::layer_data(p)$fill
+  p2 <- p + scale_fill_hypso_d()
+
+  mod <- ggplot2::layer_data(p2)$fill
+  expect_true(!any(init %in% mod))
+
+  # Alpha
+  expect_error(p + scale_fill_hypso_d(alpha = -1),
+    regexp = "alpha level -1 not in"
+  )
+
+  p3 <- p + scale_fill_hypso_d(alpha = 0.9)
+
+  mod_alpha <- ggplot2::layer_data(p3)$fill
+
+  expect_true(all(adjustcolor(mod, alpha.f = 0.9) == mod_alpha))
+
+  # Reverse also with alpha
+  expect_error(p + scale_fill_hypso_d(direction = 0.5),
+    regexp = "must be 1 or -1"
+  )
+
+
+
+  p4 <- p + scale_fill_hypso_d(
+    direction = -1,
+    alpha = 0.7
+  )
+
+  mod_alpha_rev <- ggplot2::layer_data(p4)$fill
+
+
+  expect_true(all(rev(adjustcolor(mod,
+    alpha.f = 0.7
+  )) == mod_alpha_rev))
+})
+
 test_that("Continous scale", {
   d <- data.frame(x = 1:5, y = 1:5, z = 21:25, l = letters[1:5])
 
@@ -5,30 +48,30 @@ test_that("Continous scale", {
     ggplot2::geom_col(aes(x, y, fill = z))
 
   init <- ggplot2::layer_data(p)$fill
-  p2 <- p + scale_fill_topo_c()
+  p2 <- p + scale_fill_hypso_c()
 
   mod <- ggplot2::layer_data(p2)$fill
   expect_true(!any(init %in% mod))
 
   # Alpha
-  expect_error(p + scale_fill_topo_c(alpha = -1),
+  expect_error(p + scale_fill_hypso_c(alpha = -1),
     regexp = "alpha level -1 not in"
   )
 
-  p3 <- p + scale_fill_topo_c(alpha = 0.9)
+  p3 <- p + scale_fill_hypso_c(alpha = 0.9)
 
   mod_alpha <- ggplot2::layer_data(p3)$fill
 
   expect_true(all(adjustcolor(mod, alpha.f = 0.9) == mod_alpha))
 
   # Reverse also with alpha
-  expect_error(p + scale_fill_topo_c(direction = 0.5),
+  expect_error(p + scale_fill_hypso_c(direction = 0.5),
     regexp = "must be 1 or -1"
   )
 
 
 
-  p4 <- p + scale_fill_topo_c(
+  p4 <- p + scale_fill_hypso_c(
     direction = -1,
     alpha = 0.7
   )
@@ -62,7 +105,7 @@ test_that("Breaking scale", {
 
 
   p2 <- p_init +
-    scale_fill_topo_b(breaks = br)
+    scale_fill_hypso_b(breaks = br)
 
   mod <- ggplot2::layer_data(p2)$fill
   expect_true(!any(init %in% mod))
@@ -70,11 +113,11 @@ test_that("Breaking scale", {
   expect_true(length(unique(mod)) == 3)
 
   # Alpha
-  expect_error(p_init + scale_fill_topo_b(alpha = -1),
+  expect_error(p_init + scale_fill_hypso_b(alpha = -1),
     regexp = "alpha level -1 not in"
   )
 
-  p3 <- p_init + scale_fill_topo_b(
+  p3 <- p_init + scale_fill_hypso_b(
     alpha = 0.9,
     breaks = br
   )
@@ -85,13 +128,13 @@ test_that("Breaking scale", {
   expect_true(length(unique(mod_alpha)) == 3)
 
   # Reverse also with alpha
-  expect_error(p + scale_fill_topo_b(direction = 0.5),
+  expect_error(p + scale_fill_hypso_b(direction = 0.5),
     regexp = "must be 1 or -1"
   )
 
 
 
-  p4 <- p_init + scale_fill_topo_b(
+  p4 <- p_init + scale_fill_hypso_b(
     direction = -1,
     alpha = 0.7,
     breaks = br
@@ -99,4 +142,12 @@ test_that("Breaking scale", {
 
   mod_alpha_rev <- ggplot2::layer_data(p4)$fill
   expect_true(length(unique(mod_alpha_rev)) == 3)
+})
+
+test_that("Palette", {
+  a <- hypso.colors(10)
+  expect_length(a, 10)
+
+  # NULL
+  expect_identical(hypso.colors(0), character())
 })