+++ NEWS.md
@@ -11,7 +11,7 @@
 -   Fix errors on `slice_min()` and `slice_max()` for SpatRasters (#20). Also
     add a new parameter `na.rm`.
 
--   Add new scale: `scale_fill_topo_c()`
+-   Add new scale: `scale_fill_hypso_c()`
 
 # tidyterra 0.1.0